requirements.txt
# Python dependencies for Sponsor Finder

# OpenAI API client (primary)
openai>=1.0.0

# Anthropic API client (fallback)
anthropic>=0.8.0

# Environment variable management
python-dotenv>=1.0.0

# Web scraping and research (for web-researcher agent)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data handling
pandas>=2.0.0

# Salesforce CRM integration
simple-salesforce>=1.12.0

# Optional: For enhanced logging
colorama>=0.4.6
